+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swap/info/SwapInfoFragment.kt
@@ -1,7 +1,5 @@
 package io.horizontalsystems.bankwallet.modules.swap.info
 
-import android.content.Intent
-import android.net.Uri
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -9,6 +7,7 @@ import android.view.ViewGroup
 import androidx.fragment.app.viewModels
 import io.horizontalsystems.bankwallet.R
 import io.horizontalsystems.bankwallet.core.BaseFragment
+import io.horizontalsystems.bankwallet.ui.helpers.LinkHelper
 import io.horizontalsystems.core.findNavController
 import kotlinx.android.synthetic.main.fragment_swap_info.*
 
@@ -40,9 +39,9 @@ class SwapInfoFragment : BaseFragment() {
         btnLink.text = viewModel.linkText
 
         btnLink.setOnClickListener {
-            val intent = Intent(Intent.ACTION_VIEW)
-            intent.data = Uri.parse(viewModel.dexUrl)
-            startActivity(intent)
+            context?.let { ctx ->
+                LinkHelper.openLinkInAppBrowser(ctx, viewModel.dexUrl)
+            }
         }
     }
